move prompt embed stuff
